+++ _monorepo_npm__dependencies__update__cli/index.js
@@ -1,8 +1,8 @@
 /// <reference types="./index.d.ts" />
 import { param_r_ } from 'ctx-core/cli-args'
 import { entries_gen_ } from 'ctx-core/object'
-import { createWriteStream } from 'fs'
-import { readFile } from 'fs/promises'
+import { createWriteStream } from 'node:fs'
+import { readFile } from 'node:fs/promises'
 /**
  * @param {_monorepo_npm__dependencies__update__cli__params_T}params
  * @returns {Promise<null>}