+++ monorepo_npm__version__sync__cli/index.js
@@ -1,7 +1,7 @@
 import { param_r_ } from 'ctx-core/cli-args'
 import { entries_gen_ } from 'ctx-core/object'
-import { createWriteStream } from 'fs'
-import { readFile } from 'fs/promises'
+import { createWriteStream } from 'node:fs'
+import { readFile } from 'node:fs/promises'
 import { monorepo_npm__version__sync } from '../monorepo_npm__version__sync/index.js'
 /**
  * @returns {Promise<void>}