+++ monorepo_pnpm__dependencies__update/index.js
@@ -1,6 +1,6 @@
 /// <reference types="../types/index.d.ts" />
 import { line__transform_stream_ } from 'ctx-core/string'
-import { spawn } from 'child_process'
+import { spawn } from 'node:child_process'
 import { Readable } from 'stream'
 import { monorepo_npm__dependencies__update } from '../monorepo_npm__dependencies__update/index.js'
 /**